# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Lista Ordenes registradas por los usuarios clientes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Recupera un listado Ordenes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Se crean ordenes con status en Revision, se crea el detalle valida Autorizacion del usuario;Que no exista orden ya creada (status R); Que exista el articulo y que este disponible para crear el detalle; Que exista el codigo de promocion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order",
                    "Promotion"
                ],
                "summary": "Crea registro de las ordenes de los",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "orderdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderDTOCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageInfo"
                        }
                    }
                }
            }
        },
...
        "/api/v1/orders/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Obtiene informacion de la orden segun su ID Valida orden a mostrar asociada a usuario de session para el rol cliente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Recupera una Orden segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "modifica la Orden (promocion y status) segun su ID valida Autorizacion del usuario; Que el status de entrada sea R/C; Que exista la orden a editar; Que la orden pertenezca al usuario (cliente); Que exista el codigo de promocion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Modifica una Orden segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Edit order",
                        "name": "orderdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderDTOUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Elimina una orden y su detalla, libera disponibilidad de articulos y promociones asociadas. Valida: 1.- que el usuario este autorizado;2.-Que exista la orden registrada. 3.-Que no se elimine la orden si no pertenece al usuario (cliente).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Elimina una Orden segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.MessageInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "dto.OrderDTOCreate": {
            "type": "object",
            "required": [
                "item_id",
                "quantity"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "item_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dto.OrderDTOUpdate": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.PromotionDTOCreate": {
            "type": "object",
            "required": [
                "code",
                "discount",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "discount": {
                    "type": "number",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
...
        "dto.UserDTOOrder": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "detail_order": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItems"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "order_number": {
                    "type": "integer"
                },
                "promotion": {
                    "$ref": "#/definitions/dto.PromotionDTOCreate"
                },
                "promotion_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "total_discount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.UserDTOOrder"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.OrderItems": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "name_item": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}